+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
+- Use internal coordinator communication for cursors if specified coordinator was not found on endpoint list
 
 ## [1.0.0](https://github.com/arangodb/go-driver/tree/1.0.0) (N/A)
 - Enable proper CHANGELOG and versioning